validator test cases dded
